Fix: Rename useGhostTextRenderer.js to .jsx for Vite compatibility
The file `src/hooks/useGhostTextRenderer.js` contained JSX syntax but had a .js extension. This caused issues with Vite's build analysis during Vercel deployment.

This commit renames the file to `src/hooks/useGhostTextRenderer.jsx` and updates all corresponding import statements in `PaperDisplay.jsx` and `useGhostTextRenderer.test.js` to use the new .jsx extension.

This change resolves the Vite build error related to invalid JS syntax in the hook file.
