+++ src/hooks/useGhostTextRenderer.jsx
